[BE] fix: entity 자료형 수정

+++ backend/src/main/java/com/danim/repository/NationRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface NationRepository extends JpaRepository<Nation, Long> {
+public interface NationRepository extends JpaRepository<Nation, Integer> {
     Nation findFirstByName(String address1);
 
 }
